Implement modules with clean architecture

+++ app/src/main/java/com/example/secondmvvmtrainingproject/dao/PokemonEntity.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "PokemonEntity")
-class PokemonEntity(@PrimaryKey(autoGenerate = false) val id: Long,
+class PokemonEntity(@PrimaryKey(autoGenerate = false) val id: Int,
                     val name: String,
                     val img: String,
                     val weight: String,
